# services/design-advisory/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FATHOM -- Design Advisory (Redesign Case Builder demo)",
    "version": "0.1"
  },
  "paths": {
...
    "/api/v1/design-advisory/redesign-candidates": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_api_v1_design_advisory_redesign_candidates_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Candidates Api V1 Design Advisory Redesign Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/redesign-candidates/{candidate_id}": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "Get Candidate",
        "operationId": "get_candidate_api_v1_design_advisory_redesign_candidates__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Candidate Api V1 Design Advisory Redesign Candidates  Candidate Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/dossiers/{dossier_id}": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "Get Dossier",
        "operationId": "get_dossier_api_v1_design_advisory_dossiers__dossier_id__get",
        "parameters": [
          {
            "name": "dossier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dossier Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dossier Api V1 Design Advisory Dossiers  Dossier Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/impact-snapshots": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "List Impact Snapshots",
        "operationId": "list_impact_snapshots_api_v1_design_advisory_impact_snapshots_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Impact Snapshots Api V1 Design Advisory Impact Snapshots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/impact-snapshots/{impact_snapshot_id}": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "Get Impact Snapshot",
        "operationId": "get_impact_snapshot_api_v1_design_advisory_impact_snapshots__impact_snapshot_id__get",
        "parameters": [
          {
            "name": "impact_snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Impact Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Impact Snapshot Api V1 Design Advisory Impact Snapshots  Impact Snapshot Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/cost-estimates": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "List Cost Estimates",
        "operationId": "list_cost_estimates_api_v1_design_advisory_cost_estimates_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Method"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Cost Estimates Api V1 Design Advisory Cost Estimates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/cost-estimates/{estimate_id}": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "Get Cost Estimate",
        "operationId": "get_cost_estimate_api_v1_design_advisory_cost_estimates__estimate_id__get",
        "parameters": [
          {
            "name": "estimate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Estimate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Cost Estimate Api V1 Design Advisory Cost Estimates  Estimate Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/gate-decisions": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "List Gate Decisions",
        "operationId": "list_gate_decisions_api_v1_design_advisory_gate_decisions_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Gate Decisions Api V1 Design Advisory Gate Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
    "/api/v1/design-advisory/redesign-cases": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_v1_design_advisory_redesign_cases_get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Cases Api V1 Design Advisory Redesign Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      },
      "post": {
        "tags": [
          "design-advisory:actions"
        ],
        "summary": "Create Case",
        "operationId": "create_case_api_v1_design_advisory_redesign_cases_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Case Api V1 Design Advisory Redesign Cases Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "state-changing"
      }
    },
    "/api/v1/design-advisory/redesign-cases/{case_id}": {
      "get": {
        "tags": [
          "design-advisory:reads"
        ],
        "summary": "Get Case",
        "operationId": "get_case_api_v1_design_advisory_redesign_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Case Api V1 Design Advisory Redesign Cases  Case Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-substitution": "required",
        "x-side-effects": "none"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCaseRequest": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          }
        },
        "type": "object",
        "required": [
          "candidate_id"
        ],
        "title": "CreateCaseRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}